+++ Expanded Cheat Menu/Code/Menus/CapacityMenu.lua
@@ -10,15 +10,17 @@ local c = #Actions
 -- menu
 c = c + 1
 Actions[c] = {ActionName = Translate(109035890389--[[Capacity]]),
-	ActionMenubar = "ECM.ECM",
+	ActionMenubar = "ECM.ECM.Buildings",
 	ActionId = ".Capacity",
 	ActionIcon = "CommonAssets/UI/Menu/folder.tga",
 	OnActionEffect = "popup",
+	ActionSortKey = "1Capacity",
 }
+
 
 c = c + 1
 Actions[c] = {ActionName = Strings[302535920000565--[[Storage Mechanized Depots Temp]]],
-	ActionMenubar = "ECM.ECM.Capacity",
+	ActionMenubar = "ECM.ECM.Buildings.Capacity",
 	ActionId = ".Storage Mechanized Depots Temp",
 	ActionIcon = "CommonAssets/UI/Menu/Cube.tga",
 	RolloverText = function()
@@ -32,7 +34,7 @@ Actions[c] = {ActionName = Strings[302535920000565--[[Storage Mechanized Depots 
 
 c = c + 1
 Actions[c] = {ActionName = Strings[302535920000567--[[Worker Capacity]]],
-	ActionMenubar = "ECM.ECM.Capacity",
+	ActionMenubar = "ECM.ECM.Buildings.Capacity",
 	ActionId = ".Worker Capacity",
 	ActionIcon = "CommonAssets/UI/Menu/scale_gizmo.tga",
 	RolloverText = function()
@@ -49,7 +51,7 @@ Actions[c] = {ActionName = Strings[302535920000567--[[Worker Capacity]]],
 
 c = c + 1
 Actions[c] = {ActionName = Strings[302535920000569--[[Building Capacity]]],
-	ActionMenubar = "ECM.ECM.Capacity",
+	ActionMenubar = "ECM.ECM.Buildings.Capacity",
 	ActionId = ".Building Capacity",
 	ActionIcon = "CommonAssets/UI/Menu/scale_gizmo.tga",
 	RolloverText = function()
@@ -66,7 +68,7 @@ Actions[c] = {ActionName = Strings[302535920000569--[[Building Capacity]]],
 
 c = c + 1
 Actions[c] = {ActionName = Strings[302535920000571--[[Building Visitor Capacity]]],
-	ActionMenubar = "ECM.ECM.Capacity",
+	ActionMenubar = "ECM.ECM.Buildings.Capacity",
 	ActionId = ".Building Visitor Capacity",
 	ActionIcon = "CommonAssets/UI/Menu/scale_gizmo.tga",
 	RolloverText = function()
@@ -83,7 +85,7 @@ Actions[c] = {ActionName = Strings[302535920000571--[[Building Visitor Capacity]
 
 c = c + 1
 Actions[c] = {ActionName = Strings[302535920000573--[[Storage Universal Depot]]],
-	ActionMenubar = "ECM.ECM.Capacity",
+	ActionMenubar = "ECM.ECM.Buildings.Capacity",
 	ActionId = ".Storage Universal Depot",
 	ActionIcon = "CommonAssets/UI/Menu/MeasureTool.tga",
 	RolloverText = function()
@@ -98,7 +100,7 @@ Actions[c] = {ActionName = Strings[302535920000573--[[Storage Universal Depot]]]
 
 c = c + 1
 Actions[c] = {ActionName = Strings[302535920000575--[[Storage Other Depot]]],
-	ActionMenubar = "ECM.ECM.Capacity",
+	ActionMenubar = "ECM.ECM.Buildings.Capacity",
 	ActionId = ".Storage Other Depot",
 	ActionIcon = "CommonAssets/UI/Menu/MeasureTool.tga",
 	RolloverText = function()
@@ -113,7 +115,7 @@ Actions[c] = {ActionName = Strings[302535920000575--[[Storage Other Depot]]],
 
 c = c + 1
 Actions[c] = {ActionName = Strings[302535920000577--[[Storage Waste Depot]]],
-	ActionMenubar = "ECM.ECM.Capacity",
+	ActionMenubar = "ECM.ECM.Buildings.Capacity",
 	ActionId = ".Storage Waste Depot",
 	ActionIcon = "CommonAssets/UI/Menu/MeasureTool.tga",
 	RolloverText = function()
@@ -128,7 +130,7 @@ Actions[c] = {ActionName = Strings[302535920000577--[[Storage Waste Depot]]],
 
 c = c + 1
 Actions[c] = {ActionName = Strings[302535920000579--[[Storage Mechanized Depots]]],
-	ActionMenubar = "ECM.ECM.Capacity",
+	ActionMenubar = "ECM.ECM.Buildings.Capacity",
 	ActionId = ".Storage Mechanized Depots",
 	ActionIcon = "CommonAssets/UI/Menu/Cube.tga",
 	RolloverText = function()