+++ Mods ChoGGi/Hanging Gardens Children Allowed/Code/Script.lua
@@ -1,5 +1,28 @@
 -- See LICENSE for terms
 
+local options
+local mod_PlaygroundPerk
+
+-- fired when settings are changed and new/load
+local function ModOptions()
+	mod_PlaygroundPerk = options.PlaygroundPerk
+end
+
+-- load default/saved settings
+function OnMsg.ModsReloaded()
+	options = CurrentModOptions
+	ModOptions()
+end
+
+-- fired when option is changed
+function OnMsg.ApplyModOptions(id)
+	if id ~= "ChoGGi_HangingGardensChildrenAllowed" then
+		return
+	end
+
+	ModOptions()
+end
+
 local function StartupCode()
 	BuildingTemplates.HangingGardens.usable_by_children = true
 	ClassTemplates.Building.HangingGardens.usable_by_children = true
@@ -7,3 +30,12 @@ end
 
 OnMsg.CityStart = StartupCode
 OnMsg.LoadGame = StartupCode
+
+local orig_Service = Service.Service
+function Service:Service(unit, ...)
+	if mod_PlaygroundPerk and unit.age_trait == "Child" then
+		unit.playground_visit = true
+	end
+	return orig_Service(self, unit, ...)
+end
+