+++ Mods ChoGGi/Landscaping Low Priority/metadata.lua
@@ -1,0 +1,19 @@
+return PlaceObj("ModDef", {
+	"title", "Landscaping Low Priority",
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"saved", 0,
+	"image", "Preview.png",
+	"id", "ChoGGi_LandscapingLowPriority",
+	"steam_id", "1781288394",
+	"pops_any_uuid", "a358dd9a-ac9d-4ca1-9ea0-debaded3c7dc",
+	"author", "ChoGGi",
+	"lua_revision", 245618,
+	"code", {
+		"Code/Script.lua",
+	},
+	"TagGameplay", true,
+	"TagLandscaping", true,
+	"description", [[All landscaping jobs start as low priority.]],
+})